task 5.14 & functions.js deleted
