+++ js/data.js
@@ -1,0 +1,31 @@
+export const URL_PREFIX = 'photos/';
+export const URL_SUFFIX = '.jpg';
+
+export const AVATAR_PREFIX = 'img/photos';
+export const AVATAR_SUFFIX = '.svg';
+export const POST_COUNT = 25;
+
+export const DESCRIPTIONS = [
+  'Просто красивое фото.',
+  'Обожаю это фото.',
+  'Буду публиковать чаще такое!',
+  'Щас на таком.',
+  'Нет слов.',
+  'Я просто в шоке от увиденного...'
+];
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+export const NAMES = [
+  'Андрюха192',
+  'Глеб Сасавот',
+  'Твоя подружка',
+  'Иван Золо',
+  'Пользователь WhatsApp',
+  'ОКнутый юзер'
+];
