--- js/functions.js
@@ -1,17 +1,0 @@
-/* eslint-disable no-console */
-function checkLength(string, maxLength) {
-  return string.length <= maxLength;
-}
-
-function isPalindrome(string) {
-  let left = 0;
-  let right = string.length - 1;
-  while (left < right) {
-    if (string[left] !== string[right]) {
-      return false;
-    }
-    left += 1;
-    right -= 1;
-  }
-  return true;
-}
